+++ sdk/storage/azure-storage-files-datalake/CMakeLists.txt
@@ -28,8 +28,7 @@ endif()
 
 set(
   AZURE_STORAGE_FILES_DATALAKE_HEADER
-    inc/azure/storage/files/datalake/protocol/datalake_rest_client.hpp
-    inc/azure/storage/files/datalake/datalake_constants.hpp
+    inc/azure/storage/files/datalake.hpp
     inc/azure/storage/files/datalake/datalake_directory_client.hpp
     inc/azure/storage/files/datalake/datalake_file_client.hpp
     inc/azure/storage/files/datalake/datalake_file_system_client.hpp
@@ -39,9 +38,10 @@ set(
     inc/azure/storage/files/datalake/datalake_responses.hpp
     inc/azure/storage/files/datalake/datalake_sas_builder.hpp
     inc/azure/storage/files/datalake/datalake_service_client.hpp
-    inc/azure/storage/files/datalake/datalake_utilities.hpp
     inc/azure/storage/files/datalake/dll_import_export.hpp
-    inc/azure/storage/files/datalake.hpp
+    inc/azure/storage/files/datalake/internal/datalake_utilities.hpp
+    inc/azure/storage/files/datalake/internal/datalake_constants.hpp
+    inc/azure/storage/files/datalake/protocol/datalake_rest_client.hpp
 )
 
 set(
