+++ sdk/storage/azure-storage-files-datalake/inc/azure/storage/files/datalake/internal/datalake_utilities.hpp
