+++ sdk/storage/azure-storage-files-datalake/src/datalake_file_client.cpp
@@ -9,8 +9,8 @@
 #include <azure/storage/common/internal/shared_key_policy.hpp>
 #include <azure/storage/common/storage_common.hpp>
 
-#include "azure/storage/files/datalake/datalake_constants.hpp"
-#include "azure/storage/files/datalake/datalake_utilities.hpp"
+#include "azure/storage/files/datalake/internal/datalake_constants.hpp"
+#include "azure/storage/files/datalake/internal/datalake_utilities.hpp"
 
 namespace Azure { namespace Storage { namespace Files { namespace DataLake {
