+++ sdk/storage/azure-storage-files-datalake/src/datalake_file_system_client.cpp
@@ -13,11 +13,11 @@
 #include <azure/storage/common/internal/storage_switch_to_secondary_policy.hpp>
 #include <azure/storage/common/storage_common.hpp>
 
-#include "azure/storage/files/datalake/datalake_constants.hpp"
 #include "azure/storage/files/datalake/datalake_directory_client.hpp"
 #include "azure/storage/files/datalake/datalake_file_client.hpp"
 #include "azure/storage/files/datalake/datalake_path_client.hpp"
-#include "azure/storage/files/datalake/datalake_utilities.hpp"
+#include "azure/storage/files/datalake/internal/datalake_constants.hpp"
+#include "azure/storage/files/datalake/internal/datalake_utilities.hpp"
 
 #include "private/package_version.hpp"
