+++ sdk/storage/azure-storage-files-datalake/src/datalake_service_client.cpp
@@ -14,8 +14,7 @@
 #include <azure/storage/common/storage_credential.hpp>
 
 #include "azure/storage/files/datalake/datalake_file_system_client.hpp"
-#include "azure/storage/files/datalake/datalake_utilities.hpp"
-
+#include "azure/storage/files/datalake/internal/datalake_utilities.hpp"
 #include "private/package_version.hpp"
 
 namespace Azure { namespace Storage { namespace Files { namespace DataLake {
