+++ sdk/storage/azure-storage-files-datalake/src/datalake_utilities.cpp
@@ -1,11 +1,11 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // SPDX-License-Identifier: MIT
 
-#include "azure/storage/files/datalake/datalake_utilities.hpp"
+#include "azure/storage/files/datalake/internal/datalake_utilities.hpp"
 
 #include <azure/storage/common/crypt.hpp>
 
-#include "azure/storage/files/datalake/datalake_constants.hpp"
+#include "azure/storage/files/datalake/internal/datalake_constants.hpp"
 #include "azure/storage/files/datalake/protocol/datalake_rest_client.hpp"
 
 namespace Azure { namespace Storage { namespace Files { namespace DataLake { namespace _detail {
